Fixed State Debt Resolution Bug, Increased Contract Distance Buffer to 10
Fixed State Debt Resolution Bug - contractDestinationCompany and shipperCompany parameters were around the wrong way

Increased Contract Distance Buffer to 10 - Trucks were missing delivery windows too often and their companies were getting penalised too heavily. 10 might be a bit extreme though.. maybe 6-8 instead?

+++ unity-client/Assets/Scripts/DTO/CompanyDTO.cs
@@ -1,5 +1,5 @@
 public class CompanyDTO : BaseEntityDTO, INamedEntity
 {
     public string Name { get; set; }
-    public int Money { get; set; }
+    public float Money { get; set; }
 }
